Split monolithic MCP Server
